feat(notifications): rename component
