+++ src/components/profile/ProfileCard/ProfileFormNav/nav/Notifications/NotificationsTypes/CheckboxColumn/CheckboxColumn.styles.ts
