+++ src/components/profile/ProfileCard/ProfileFormNav/nav/Notifications/NotificationsTypes/CheckboxColumn/CheckboxColumn.tsx
@@ -3,15 +3,15 @@ import { Checkbox } from 'antd';
 import { Option } from '../interfaces';
 import { CheckboxChangeEvent } from 'antd/lib/checkbox';
 import { CheckboxValueType } from 'antd/lib/checkbox/Group';
-import * as S from './NotificationGroup.styles';
+import * as S from './CheckboxColumn.styles';
 
-interface NotificationGroupProps {
+interface CheckboxColumnProps {
   column: Omit<Option, 'id'>;
   handleCheck: (list: CheckboxValueType[]) => void;
   setTriggered: (state: boolean) => void;
 }
 
-export const NotificationGroup: React.FC<NotificationGroupProps> = ({ column, handleCheck, setTriggered }) => {
+export const CheckboxColumn: React.FC<CheckboxColumnProps> = ({ column, handleCheck, setTriggered }) => {
   const [checkedList, setCheckedList] = React.useState<CheckboxValueType[]>([]);
   const [indeterminate, setIndeterminate] = React.useState(false);
   const [checkAll, setCheckAll] = React.useState(false);
