+++ src/components/profile/ProfileCard/ProfileFormNav/nav/Notifications/NotificationsTypes/NotificationsTypes.tsx
@@ -3,7 +3,7 @@ import { Checkbox } from 'antd';
 import { useTranslation } from 'react-i18next';
 import { ProfileForm } from '../../../ProfileForm/ProfileForm';
 import { Option } from './interfaces';
-import { NotificationGroup } from './NotificationGroup/NotificationGroup';
+import { CheckboxColumn } from './CheckboxColumn/CheckboxColumn';
 import { CheckboxValueType } from 'antd/lib/checkbox/Group';
 import * as S from './NotificationsTypes.styles';
 
@@ -81,12 +81,7 @@ export const NotificationsTypes: React.FC = () => {
     <ProfileForm name="notifications" onCancel={onCancel} onFinish={onFinish} trigger={isTriggered}>
       <S.Wrapper>
         {options.map((item) => (
-          <NotificationGroup
-            key={item.id}
-            column={item}
-            handleCheck={handleCheck(item.id)}
-            setTriggered={setTriggered}
-          />
+          <CheckboxColumn key={item.id} column={item} handleCheck={handleCheck(item.id)} setTriggered={setTriggered} />
         ))}
       </S.Wrapper>
     </ProfileForm>
